type migration to handle bomb explosion

+++ core/src/main/java/se2/ticktackbumm/core/network/messages/client/BombStart.java
@@ -2,17 +2,17 @@ package se2.ticktackbumm.core.network.messages.client;
 
 public class BombStart {
 
-    private  int bombTimer;
+    private float bombTimer;
 
     public BombStart(){
 
     }
 
-    public BombStart(int bombTimer){
+    public BombStart(float bombTimer){
         this.bombTimer = bombTimer;
     }
 
-    public int getBombTimer(){
+    public float getBombTimer(){
         return this.bombTimer;
     }
